requirements.txt
"""
Listado de dependencias para entornos alternativos donde no se use UV directamente.
Utilizado por herramientas como Docker para la instalación con pip.
"""

# Core YOLO/Deep Learning ---
ultralytics==8.3.205        # YOLOv8 oficial
torch==2.8.0
torchvision==0.23.0
ultralytics-thop==2.0.17

# NumPy/Scientific stack ---
numpy==2.2.6
scipy==1.16.2
pandas==2.3.3
pyarrow==21.0.0
polars==1.34.0
psutil==7.1.0

# CV / Visualización ---
opencv-python==4.12.0.88    # Procesamiento de imagen y video
matplotlib==3.10.6
pillow==11.3.0              # Manejo de imágenes

# Utilidades varias ---
pyyaml==6.0.3
requests==2.32.5

# Comunicación con gRPC ---
protobuf==6.32.1
grpcio==1.75.1 
grpcio-tools==1.75.1

# MLflow / tracking
mlflow==2.16.2                   

# Streamlit (web demo
streamlit>=1.50.0           
altair==5.5.0
pydeck==0.9.1
tenacity==9.1.2
toml==0.10.2
watchdog==6.0.0
tornado==6.5.2

# Subset script / utilidades de datos ---
scikit-learn==1.5.2

# pruebas unitarias ---
pytest>=7.4.0              # 



